+++ src/avatar/animations/loader/index.ts
@@ -4,7 +4,7 @@ import {
   type GLTF,
 } from 'three/examples/jsm/loaders/GLTFLoader.js';
 import { Logger } from '../../../logger.js';
-import { AvatarModelConfig, GestureMappingKeys } from '../../webavatar.dto.js';
+import { AvatarModelConfig, GestureMappingKeys } from 'dto.js';
 import { RpmAnimation } from './contants.js';
 
 export class AnimationsLoader {
