+++ src/avatar/index.ts
@@ -1,3 +1,3 @@
 export { LipSync } from './lipsync/index.js';
-export * from './webavatar.dto.js';
+export * from '../dto/webavatar.dto.js';
 export * from './webavatar.js';
