--- src/detection/detection.dto.ts
@@ -1,6 +1,0 @@
-export type WorkerMessageType = 'init' | 'process' | 'render' | 'destroy';
-
-export interface WorkerMessage extends Record<string, any> {
-  type: WorkerMessageType;
-  data: any;
-}
