+++ src/detection/video/base.detector.ts
@@ -1,5 +1,5 @@
 import { Logger } from '../../logger.js';
-import { type WorkerMessage } from '../detection.dto.js';
+import { type WorkerMessage } from 'dto.js';
 import { VideoDetector, VideoDetectorConfig } from './video.dto.js';
 
 export abstract class BaseDetector<
