+++ src/dto.ts
@@ -8,3 +8,5 @@ export * from './dto/proxy.dto.js';
 export * from './dto/session.dto.js';
 export * from './dto/settings.dto.js';
 export * from './dto/ui.dto.js';
+export * from './dto/webavatar.dto.js';
+export * from './dto/webavatar.audio-player.dto.js';
