+++ src/dto/detection.dto.ts
@@ -21,3 +21,10 @@ export enum UserCharacterizationEventSource {
   emotion_tracker = 'emotion_tracker',
   speechbrain = 'speechbrain',
 }
+
+export type WorkerMessageType = 'init' | 'process' | 'render' | 'destroy';
+
+export interface WorkerMessage extends Record<string, any> {
+  type: WorkerMessageType;
+  data: any;
+}
